git-svn-id: svn+ssh://src.apple.com/svn/cups/easysw/current@1009 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ scheduler/select.c
@@ -638,7 +638,7 @@ cupsdDoSelect(long timeout)		/* I - Timeout in seconds */
     {
       cupsdLogMessage(CUPSD_LOG_DEBUG2,
                       "cupsdDoSelect: pollfds[%d]={fd=%d, revents=%x}",
-		      pfd - cupsd_pollfds, pfd->fd, pfd->revents);
+		      (int)(pfd - cupsd_pollfds), pfd->fd, pfd->revents);
 
       if (!pfd->revents)
         continue;
